Move settings policy from governance to propose endpoint
