+++ test/system-test/test_settingsPolicy.py
@@ -8,11 +8,6 @@ from utils import (
 from endpoints import settingsPolicy
 
 
-@pytest.mark.xfail(
-    os.getenv("TEST_ENVIRONMENT") == "ccf/acl",
-    strict=True,
-    reason="Governance operations need to move to user endpoints",
-)
 def test_settingsPolicy_with_no_policy(setup_kms):
     status_code, settings_json = settingsPolicy()
     assert status_code == 200
@@ -31,11 +26,6 @@ def test_settingsPolicy_with_no_auth(setup_kms):
     assert status_code == 401
 
 
-@pytest.mark.xfail(
-    os.getenv("TEST_ENVIRONMENT") == "ccf/acl",
-    strict=True,
-    reason="Governance operations need to move to user endpoints",
-)
 def test_settingsPolicy_with_policy(setup_kms):
     apply_kms_constitution()
 
@@ -54,11 +44,6 @@ def test_settingsPolicy_with_policy(setup_kms):
     assert settings_json == policy
 
 
-@pytest.mark.xfail(
-    os.getenv("TEST_ENVIRONMENT") == "ccf/acl",
-    strict=True,
-    reason="Governance operations need to move to user endpoints",
-)
 def test_settingsPolicy_with_multiple_policy_sets(setup_kms):
     apply_kms_constitution()
